chore: extract lib/router.ts file
